user rate for a movie ok

+++ dtos/RateResDTO.go
@@ -2,5 +2,5 @@ package dtos
 
 type RateResDTO struct {
 	MovieDbId int32 `bson:"movie_db_id" json:"movie_db_id"`
-	Rate      int8  `bson:"rate,omitempty" json:"rate"`
+	Rate      float32  `bson:"rate,omitempty" json:"rate"`
 }
